fix: category status bar overspent render
The category status bar was not being properly drawn as there were only
2 possible states for the "base bar", the one that draws the base color
of the status: funded or not. The truth is that there could be a third
state: `overspent`. This change addresses the fact by applying green
when the category is either `funded` or `in_spending` and red when its
overspent. In any other case, the default gray will do the work.